Re-structured folders based on architecture
