+++ express/app.js
@@ -4,14 +4,14 @@ var path = require("path");
 var cookieParser = require("cookie-parser");
 var logger = require("morgan");
 
-var indexRouter = require("./apis");
-var usersRouter = require("./apis/users");
+var indexRouter = require("./apiServices");
+var usersRouter = require("./apiServices/users");
 
 var app = express();
 
 app.use(logger("dev"));
 app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
+app.use(express.urlencoded({extended: false}));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
